requirements.txt
email_validator==2.2.0
Flask==3.0.0
Flask-Mail==0.10.0
Flask-Uploads==0.2.1
python-dotenv==1.0.0
regex==2024.7.24
gunicorn